Adapted bullet spawning
Adapted bullet spawning to fit the new way of input handling

+++ core/src/com/mygdx/shapewars/model/components/VelocityComponent.java
@@ -6,7 +6,7 @@ public class VelocityComponent implements Component{
   private float value;
   private float direction;
 
-  public VelocityComponent(int value, int direction) {
+  public VelocityComponent(float value, float direction) {
     this.value = value;
     this.direction = direction;
   }
